Added register.php file  and  made changes  to signup
Added register.php file in order to fix the  problem that I was having
with signup but  did not need it  in the the end  after making a  change
to the  connection to the  database within the  signup. I now can add
new users to the  database when signing up. I will leave the  register
.php file for  to show  what what i was trying to do here.